test: :white_check_mark: deleted icontest because there were two and moved changes to the original
